requirements.txt

openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
tqdm>=4.66.1
pandas>=2.1.4
numpy>=1.26.2
neo4j>=5.0.0

